--- src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/annotation/package-info.java
@@ -1,3 +1,0 @@
-@Deprecated
-package com.puppycrawl.tools.checkstyle.checks.annotation;
-
