+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/annotation/packageannotation/InputPackageAnnotation2.java
@@ -1,5 +1,5 @@
 @Deprecated // non-compilable annotation
-package com.puppycrawl.tools.checkstyle.checks.annotation;
+package com.puppycrawl.tools.checkstyle.checks.annotation.packageannotation;
 
 public class InputPackageAnnotation2 {
