+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/annotation/packageannotation/package-info.java
@@ -1,0 +1,3 @@
+@Deprecated
+package com.puppycrawl.tools.checkstyle.checks.annotation.packageannotation;
+
