--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/annotation/InputPackageAnnotation.java
@@ -1,6 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.annotation;
-
-@Deprecated
-public class InputPackageAnnotation {
-	
-}
