+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/annotation/packageannotation/InputPackageAnnotation.java
@@ -1,0 +1,6 @@
+package com.puppycrawl.tools.checkstyle.checks.annotation.packageannotation;
+
+@Deprecated
+public class InputPackageAnnotation {
+	
+}
